+++ src/deepfield_jobs/deepfield_jobs/gps_track.py
@@ -22,26 +22,32 @@
 
 from __future__ import absolute_import, division
 
+from itertools import product
+from collections import defaultdict
 from marv import bb
 from marv.bb import job_logger as logger
-from marv.model import Jobrun
 from bagbunker import bb_bag
 
-__version__ = '0.0.5'
+__version__ = '0.0.6'
 
-# GPS topics with descending priority: first found will be used.
-GPS_TOPICS = (
+# POSITION topics with descending priority: first found will be used.
+POSITION_TOPICS = (
     '/sensor/gps/leica1200/fix',
     '/sensor/gps/trimble/fix',
     '/sensor/gps/evk7ppp/fix',
+    '/gps/fix',
 )
-IMU_TOPIC = '/sensor/imu/xsens_mti/data'
-TOPICS = GPS_TOPICS + (IMU_TOPIC,)
+ORIENTATION_TOPICS = (
+    '/sensor/imu/xsens_mti/data',
+    '/gps/orientation',
+)
+
+TOPICS = POSITION_TOPICS + ORIENTATION_TOPICS
 
 
-def detail_image_view(topic):
+def detail_image_view(topic, orient_topic):
     @bb.detail()
-    @bb.image_widget(title='GPS plot {}, {}'.format(topic, IMU_TOPIC))
+    @bb.image_widget(title='Position plot {}, {}'.format(topic, orient_topic))
     def detail_image(fileset):
         jobrun = fileset.get_latest_jobrun('deepfield::gps_track')
         if not jobrun:
@@ -67,35 +73,33 @@ def detail_image_view(topic):
 
 # We register a view for each gps topic, but so far the gps job can
 # only process one topic.
-for topic in GPS_TOPICS:
-    detail_image_view(topic)
+for topic, orient_topic in product(POSITION_TOPICS, ORIENTATION_TOPICS):
+    detail_image_view(topic, orient_topic)
 
 
 @bb.job()
 @bb_bag.messages(topics=TOPICS)
 def job(fileset, messages):
-    gps_topics = [x.topic.name for x in fileset.bag.topics
-                  if x.topic.name in GPS_TOPICS]
-    if not gps_topics:
+    position_topics = [x.topic.name for x in fileset.bag.topics
+                       if x.topic.name in POSITION_TOPICS]
+    orientation_topics = [x.topic.name for x in fileset.bag.topics
+                          if x.topic.name in ORIENTATION_TOPICS]
+    if not position_topics:
         logger.debug('no gps topic found')
         return []
-    for gps_topic in gps_topics:
-        logger.info('starting with {} and {}'.format(gps_topic, IMU_TOPIC))
+    logger.info('starting with {} and {}'.format(position_topics,
+                                                 orientation_topics))
 
     import datetime
     import pyproj
+    import matplotlib.dates as md
     import matplotlib.pyplot as plt
-    from matplotlib import cm, dates
+    from matplotlib import cm
     import numpy as np
-
-    def plot_convert_time(timestamps):
-        dts = map(lambda e: datetime.datetime.fromtimestamp(e),
-                  timestamps.astype(int))
-        return dates.date2num(dts)
 
     proj = pyproj.Proj(proj='utm', zone=32, ellps='WGS84')
 
-    class GPS(object):
+    class Position(object):
         def __init__(self):
             self.e_offset = 0
             self.n_offset = 0
@@ -122,87 +126,93 @@ def job(fileset, messages):
                 np.sqrt(msg.position_covariance[0])
             ])
 
-    class IMU(object):
+    class Orientation(object):
         def __init__(self):
             self.orientation = []
 
         def update(self, msg):
-            self.orientation.append([
-                msg.header.stamp.to_sec(),
-                msg.orientation.x,
-                msg.orientation.y,
-                msg.orientation.z,
-                msg.orientation.w
-            ])
+            if hasattr(msg, 'yaw') and not np.isnan(msg.yaw):
+                self.orientation.append([
+                    msg.header.stamp.to_sec(),
+                    msg.yaw
+                    ])
+            elif hasattr(msg, 'orientation') and not np.isnan(msg.orientation.x):
+                self.orientation.append([
+                    msg.header.stamp.to_sec(),
+                    self.yaw_angle(msg.orientation)
+                ])
 
-    gpsMap = dict([(gps_topic, GPS()) for gps_topic in gps_topics])
-    imu = IMU()
-    erroneous_msg_count = 0
+        # calculate imu orientation
+        @staticmethod
+        def yaw_angle(frame):
+            rot = np.zeros((3, 3))
+
+            # consists of time, x, y, z, w
+            q1 = frame.x
+            q2 = frame.y
+            q3 = frame.z
+            q4 = frame.w
+
+            rot[0, 0] = 1 - 2 * q2 * q2 - 2 * q3 * q3
+            rot[0, 1] = 2 * (q1 * q2 - q3 * q4)
+            rot[0, 2] = 2 * (q1 * q3 + q2 * q4)
+            rot[1, 0] = 2 * (q1 * q2 + q3 * q4)
+            rot[1, 1] = 1 - 2 * q1 * q1 - 2 * q3 * q3
+            rot[1, 2] = 2 * (q2 * q3 - q1 * q4)
+            rot[2, 0] = 2 * (q1 * q3 - q2 * q4)
+            rot[2, 1] = 2 * (q1 * q4 + q2 * q3)
+            rot[2, 2] = 1 - 2 * q1 * q1 - 2 * q2 * q2
+
+            vec = np.dot(rot, [1, 0, 0])
+
+            # calculate the angle
+            return np.arctan2(vec[1], vec[0])
+
+    positionMap = {position_topic: Position() for position_topic in position_topics}
+    orientationMap = {orientation_topic: Orientation()
+                      for orientation_topic in orientation_topics}
+    erroneous_msg_count = defaultdict(int)
 
     for topic, msg, timestamp in messages:
-        if topic in gps_topics:
+        if topic in position_topics:
             # skip erroneous messages
             if np.isnan(msg.longitude) or \
                np.isnan(msg.latitude) or \
                np.isnan(msg.altitude):
-                erroneous_msg_count += 1
+                erroneous_msg_count[topic] += 1
                 continue
 
             if hasattr(msg, 'status'):
-                gpsMap[topic].update(msg)
+                positionMap[topic].update(msg)
             else:
-                raise Exception('Invalid GPS topic')
-        elif topic == IMU_TOPIC:
-            imu.update(msg)
+                raise Exception('Invalid position topic')
+        elif topic in orientation_topics:
+            orientationMap[topic].update(msg)
 
     if erroneous_msg_count:
-        logger.warn('Skipped %s erroneous gps messages on topic %s',
-                    erroneous_msg_count, gps_topic)
+        logger.warn('Skipped erroneous GNSS messages %r', erroneous_msg_count.items())
 
-    # calculate imu orientation
-    def yaw_angle(frame):
-        rot = np.zeros((3, 3))
-
-        # consists of time, x, y, z, w
-        q1 = frame[1]
-        q2 = frame[2]
-        q3 = frame[3]
-        q4 = frame[4]
-
-        rot[0, 0] = 1 - 2 * q2 * q2 - 2 * q3 * q3
-        rot[0, 1] = 2 * (q1 * q2 - q3 * q4)
-        rot[0, 2] = 2 * (q1 * q3 + q2 * q4)
-        rot[1, 0] = 2 * (q1 * q2 + q3 * q4)
-        rot[1, 1] = 1 - 2 * q1 * q1 - 2 * q3 * q3
-        rot[1, 2] = 2 * (q2 * q3 - q1 * q4)
-        rot[2, 0] = 2 * (q1 * q3 - q2 * q4)
-        rot[2, 1] = 2 * (q1 * q4 + q2 * q3)
-        rot[2, 2] = 1 - 2 * q1 * q1 - 2 * q2 * q2
-
-        vec = np.dot(rot, [1, 0, 0])
-
-        # calculate the angle
-        return np.arctan2(vec[1], vec[0])
-
-    for gps_topic in gps_topics:
-        gps = np.array(gpsMap[gps_topic].gps)
+    for position_topic, orientation_topic in product(
+            position_topics, orientation_topics
+    ):
+        gps = np.array(positionMap[position_topic].gps)
 
         if not len(gps):
-            logger.error('Aborting due to missing gps messages on topic %s', gps_topic)
+            logger.error('Aborting due to missing gps messages on topic %s',
+                         position_topic)
             continue
 
-        if imu.orientation:
-            orientation = np.array(imu.orientation)
+        if orientationMap[orientation_topic].orientation:
+            orientation = np.array(orientationMap[orientation_topic].orientation)
         else:
-            logger.warn('No imu messages on topic %s', IMU_TOPIC)
-            continue
+            logger.warn('No orientation messages on topic %s', orientation_topic)
 
         # plotting
         fig = plt.figure()
         fig.subplots_adjust(wspace=0.3)
 
         ax1 = fig.add_subplot(1, 3, 1)  # e-n plot
-        ax2 = fig.add_subplot(2, 3, 2)  # imu plot
+        ax2 = fig.add_subplot(2, 3, 2)  # orientation plot
         ax3 = fig.add_subplot(2, 3, 3)  # e-time plot
         ax4 = fig.add_subplot(2, 3, 5)  # up plot
         ax5 = fig.add_subplot(2, 3, 6)  # n-time plot
@@ -210,54 +220,44 @@ def job(fileset, messages):
         # masking for finite values
         gps = gps[np.isfinite(gps[:, 1])]
 
-        # timeseries for x-axis
-        timeseries = plot_convert_time(gps[:, 0])
-
-        if imu.orientation:
-            timeseries_yaw = plot_convert_time(orientation[:, 0])
-
-            imu_yaw = map(yaw_angle, orientation)
-
         # precompute plot vars
         c = cm.prism(gps[:, 7]/2)
 
         ax1.scatter(gps[:, 4], gps[:, 5], c=c, edgecolor='none', s=3,
                     label="green: RTK\nyellow: DGPS\nred: Single")
 
-        if imu.orientation:
-            ax2.scatter(timeseries_yaw, imu_yaw, edgecolor='none', s=3)
-            ax2.set_xlim([timeseries_yaw.min(), timeseries_yaw.max()])
-            ax2.xaxis.set_major_formatter(dates.DateFormatter('%H:%M:%S'))
+        xfmt = md.DateFormatter('%H:%M:%S')
+        ax2.xaxis.set_major_formatter(xfmt)
+        ax3.xaxis.set_major_formatter(xfmt)
+        ax4.xaxis.set_major_formatter(xfmt)
+        ax5.xaxis.set_major_formatter(xfmt)
 
-        ax3.scatter(timeseries, gps[:, 4], c=c, edgecolor='none', s=3)
-        ax3.set_xlim([timeseries.min(), timeseries.max()])
-        ax3.xaxis.set_major_formatter(dates.DateFormatter('%H:%M:%S'))
+        if orientationMap[orientation_topic].orientation:
+            ax2.plot([datetime.datetime.fromtimestamp(timestamp)
+                      for timestamp in orientation[:, 0]], orientation[:, 1])
 
-        ax4.scatter(timeseries, gps[:, 6], c=c, edgecolor='none', s=3)
-        ax4.set_xlim([timeseries.min(), timeseries.max()])
-        ax4.xaxis.set_major_formatter(dates.DateFormatter('%H:%M:%S'))
+        ax3.plot([datetime.datetime.fromtimestamp(timestamp)
+                  for timestamp in gps[:, 0]], gps[:, 4])
+        ax4.plot([datetime.datetime.fromtimestamp(timestamp)
+                  for timestamp in gps[:, 0]], gps[:, 6])
+        ax5.plot([datetime.datetime.fromtimestamp(timestamp)
+                  for timestamp in gps[:, 0]], gps[:, 5])
 
-        ax5.scatter(timeseries, gps[:, 5], c=c, edgecolor='none', s=3)
-        ax5.set_xlim([timeseries.min(), timeseries.max()])
-        ax5.xaxis.set_major_formatter(dates.DateFormatter('%H:%M:%S'))
-        ax2.xaxis_date()
-        ax3.xaxis_date()
-        ax4.xaxis_date()
-        ax5.xaxis_date()
         fig.autofmt_xdate()
 
         # add the legends
         ax1.legend(loc="best")
 
-        ax1.set_ylabel('GPS northing [m]')
-        ax1.set_xlabel('GPS easting [m]')
-        ax2.set_ylabel('IMU heading over time [rad]')
-        ax3.set_ylabel('GPS easting over time [m]')
-        ax4.set_ylabel('GPS height over time [m]')
-        ax5.set_ylabel('GPS northing over time [m]')
+        ax1.set_ylabel('GNSS northing [m]')
+        ax1.set_xlabel('GNSS easting [m]')
+        ax2.set_ylabel('Heading over time [rad]')
+        ax3.set_ylabel('GNSS easting over time [m]')
+        ax4.set_ylabel('GNSS height over time [m]')
+        ax5.set_ylabel('GNSS northing over time [m]')
 
         fig.set_size_inches(16, 9)
-        path = bb.make_job_file(gps_topic.replace("/", "_")+'.jpg')
+        path = bb.make_job_file(position_topic.replace("/", "_")+'.jpg')
+
         try:
             fig.savefig(path)
         except: